Move ViewNameInferrer out of Action namespace
This avoids us awkwardly inheriting from Hanami::Action earlier than we'd otherwise intend due to Hanami::Application::Action being this class' (previous) namespace.
